requirements.txt
certifi==2021.10.8
chainer==7.8.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.4.1
filelock==3.6.0
fonttools==4.29.1
gitdb==4.0.9
gitpython==3.1.27
idna==3.3
importlib-metadata==4.11.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.3.2
lmdb==1.3.0
lpips==0.1.4
matplotlib==3.5.1
ninja==1.10.2.3
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
pillow==9.0.1
pip==21.2.2
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytz==2021.3
pyyaml==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.6
setproctitle==1.2.2
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch-fidelity==0.3.0
torch==1.8.1+cu101
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.9.1+cu101
tornado==6.1
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
visdom==0.1.8.9
wandb==0.12.11
websocket-client==1.3.1
wheel==0.37.1
yaspin==2.1.0
zipp==3.7.0